implement document/get-by-keys and change signature of document/delete-by-keys
The get-by-keys method can return several documents by taking
a vector of keys and using the /simple endpoint to look them up.

The delete-by-keys method does the same with deleting documents.
It was already implemented, but was changed in a way to conform
with the other methods better. It now takes the collection name
optionally and uses the default collection otherwise.